dom/tests/mochitest/general/test_bug629535.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=629535
-->
<head>
  <title>Test for Bug 629535</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=629535">Mozilla Bug 629535</a>

<script type="application/javascript">

const dntPref = 'privacy.donottrackheader.enabled';
const oldDNT = SpecialPowers.getBoolPref(dntPref);

is(SpecialPowers.getBoolPref(dntPref), false,
   'DNT should be disabled by default');
is(navigator.doNotTrack, 'unspecified',
   'navigator.doNotTrack should initially be "unspecified".');

SpecialPowers.clearUserPref(dntPref);
is(navigator.doNotTrack, "unspecified", 'after clearing pref');

SpecialPowers.setBoolPref(dntPref, true);
is(navigator.doNotTrack, "yes", 'after setting pref to true');

SpecialPowers.setBoolPref(dntPref, false);
is(navigator.doNotTrack, "unspecified", 'after setting pref to false');

SpecialPowers.setBoolPref(dntPref, oldDNT);

</script>

</body>
</html>
